+++ src/components/PhaseBar/PhaseBarSkeleton.tsx
