+++ src/components/PhaseBar/index.tsx
