+++ src/components/PhaseBar/styles.ts
