+++ src/hooks/instance.ts
@@ -1,6 +1,7 @@
 import { localStorageGet, localStorageSet } from '@/utils';
 import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
+import { useIsOnline } from './auth';
 
 export const useCodeManagement = () => {
   const location = useLocation();
@@ -20,5 +21,22 @@ export const useCodeManagement = () => {
   return {
     code,
     resetCode,
+  };
+};
+
+export const useIsOnlineState = () => {
+  const location = useLocation();
+  const [online, setOnline] = useState(true);
+
+  const checkOnlineStatus = async () => {
+    setOnline(await useIsOnline());
+  };
+
+  useEffect(() => {
+    checkOnlineStatus();
+  }, [location]);
+
+  return {
+    online,
   };
 };
