+++ src/layout/config.ts
@@ -8,6 +8,9 @@ import { checkPermissions } from '@/utils';
 /**
  * SideBar configuration
  */
+
+export const APP_TITLE = 'aula';
+
 const SIDEBAR_MOBILE_ANCHOR = 'right'; // 'left';
 export const SIDEBAR_DESKTOP_ANCHOR = 'right'; // 'left';
 export const SIDEBAR_WIDTH = '240px';
